third_party/blink/web_tests/platform/generic/svg/custom/svg-getelementid-expected.txt

PASS null != subElem1 is true
PASS subElem1 is subElem1
PASS null == svgTree1.getElementById('subElem2') is true
PASS null == svgTree1.getElementById('siblingElem1') is true
PASS null == svgTree3.getElementById('rect') is true
PASS null == svgTree3.getElementById('svgTree3') is true
PASS svgTree4.firstChild.nextSibling == root.getElementById('foo') is true
PASS svgTree5.firstChild.nextSibling == bar is true
PASS svgTree5.firstChild.nextSibling == svgTree5.getElementById('bar') is true
PASS foreign.firstChild.nextSibling == nonsvg is true
PASS childSvg == orphanSvg.getElementById('fooSvg2') is true
PASS successfullyParsed is true

TEST COMPLETE
